requirements.txt
selenium
playwright
pytest
requests
python-dotenv
webdriver_manager